requirement.txt
fastapi
uvicorn
onnxruntime
torch
torchvision
pillow
numpy
python-multipart